Script that allows sending custom ethernet frames
This script generates dummy pcap headers and the hex raw structure of an ethernet frame and writes the result to a file. This allows the tool Tcpreplay to interpret the resulting file as a previous data capture and send the raw data via a network interface at the speed rate of your choice.